+++ Firestore/CHANGELOG.md
@@ -24,6 +24,8 @@
   Query.getDocuments() should fetch from server only, cache only, or attempt
   server and fall back to the cache (which was the only option previously, and
   is now the default.)
+- [feature] Added new `mergeFields:(NSArray<id>*)` override for `set()`
+  which allows merging of a reduced subset of fields.
 
 # v0.11.0
 - [fixed] Fixed a regression in the Firebase iOS SDK release 4.11.0 that could